--- service/resource/vaultcrt/patch.go
@@ -1,38 +1,0 @@
-package vaultcrt
-
-import (
-	"context"
-
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/operatorkit/framework"
-)
-
-func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*framework.Patch, error) {
-	create, err := r.newCreateChange(ctx, obj, currentState, desiredState)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	update, err := r.newUpdateChange(ctx, obj, currentState, desiredState)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	patch := framework.NewPatch()
-	patch.SetCreateChange(create)
-	patch.SetUpdateChange(update)
-
-	return patch, nil
-}
-
-func (r *Resource) NewDeletePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*framework.Patch, error) {
-	delete, err := r.newDeleteChange(ctx, obj, currentState, desiredState)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	patch := framework.NewPatch()
-	patch.SetDeleteChange(delete)
-
-	return patch, nil
-}
